Add OwnedCapability
